+++ src/main/java/entity/ExchangeRate.java
@@ -3,7 +3,7 @@ package entity;
 import java.math.BigDecimal;
 
 public class ExchangeRate {
-  private Long id;
+  private Integer id;
   private Currency baseCurrency;
   private Currency targetCurrency;
   private BigDecimal rate;
@@ -11,18 +11,18 @@ public class ExchangeRate {
   public ExchangeRate() {
   }
 
-  public ExchangeRate(Long id, Currency baseCurrency, Currency targetCurrency, BigDecimal rate) {
+  public ExchangeRate(Integer id, Currency baseCurrency, Currency targetCurrency, BigDecimal rate) {
     this.id = id;
     this.baseCurrency = baseCurrency;
     this.targetCurrency = targetCurrency;
     this.rate = rate;
   }
 
-  public Long getId() {
+  public Integer getId() {
     return id;
   }
 
-  public void setId(Long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
